fix graphql circular dependencies
