+++ package.json
@@ -12,6 +12,7 @@
     "analyze-circular-dependencies-postgres": "madge packages/db-postgres/src/index.ts --circular --warning",
     "analyze-circular-dependencies-drizzle": "madge packages/drizzle/src/index.ts --circular --warning",
     "analyze-circular-dependencies-drizzle-postgres": "madge packages/drizzle/src/exports/postgres.ts --circular --warning",
+    "analyze-circular-dependencies-graphql": "madge packages/grapqhl/src/index.ts --circular --warning",
     "bf": "pnpm run build:force",
     "build": "pnpm run build:core",
     "build:all": "turbo build",
